specular highlights
also decoupled shader interface from common.h

+++ src/shader.h
@@ -1,14 +1,10 @@
 #pragma once
 
-#include "sv.h"
-#include "da.h"
-#include "common.h"
-
 #include <cglm/cglm.h>
 
-u32 Shader_CreateFromGLSLBuffer(void* buffer, u64 bufferSize);
-void Shader_Use(u32 shader);
+unsigned int Shader_CreateFromGLSLBuffer(void* buffer, unsigned long long bufferSize);
+void Shader_Use(unsigned int shader);
 
-void Shader_SetUniformFloat(u32 shader, const char* name, float value);
-void Shader_SetUniformVec3(u32 shader, const char* name, vec3 vector);
-void Shader_SetUniformMat4(u32 shader, const char* name, mat4 matrix);
+void Shader_SetUniformFloat(unsigned int shader, const char* name, float value);
+void Shader_SetUniformVec3(unsigned int shader, const char* name, vec3 vector);
+void Shader_SetUniformMat4(unsigned int shader, const char* name, mat4 matrix);
